change keyframes to float based timers
also fixes pure keyframe animations for bbmodels

+++ src/main/java/thut/api/entity/animation/Animators.java
@@ -68,7 +68,7 @@ public class Animators
         default void setColours(int... rgba)
         {}
 
-        int getLength();
+        float getLength();
 
         boolean hasLimbBased();
 
@@ -158,7 +158,7 @@ public class Animators
             }
         }
 
-        private static record AnimChannel(String channel, List<AnimationComponent> components, int length)
+        private static record AnimChannel(String channel, List<AnimationComponent> components, float length)
         {
         };
 
@@ -170,7 +170,7 @@ public class Animators
         public final List<AnimationComponent> components;
         public final List<AnimChannel> channels = new ArrayList<>();
         private final Set<CHANNEL> channelSet = new HashSet<>();
-        private int length = -1;
+        private float length = -1;
         private boolean limbBased;
         private boolean hidden = false;
 
@@ -184,7 +184,7 @@ public class Animators
             this(components, false, -1);
         }
 
-        public KeyframeAnimator(List<AnimationComponent> components, boolean preComputed, int baseLength)
+        public KeyframeAnimator(List<AnimationComponent> components, boolean preComputed, float baseLength)
         {
             this.components = components;
             this.length = -1;
@@ -275,7 +275,7 @@ public class Animators
                 var list = by_channel.get(key);
                 if (list != null)
                 {
-                    int len = 1;
+                    float len = 1;
                     for (var comp : list) len = Math.max(len, comp.startKey + comp.length);
                     AnimChannel channel = new AnimChannel(key, list, len);
                     this.channels.add(channel);
@@ -451,7 +451,7 @@ public class Animators
                 float t1 = time1, t2 = time2;
                 if (animation.loops)
                 {
-                    int l = animChannel.length();
+                    float l = animChannel.length();
                     if (l > 1)
                     {
                         t1 = time1 % l;
@@ -480,7 +480,7 @@ public class Animators
 
                 float componentTimer = time - component.startKey;
                 if (componentTimer > component.length) componentTimer = component.length;
-                final int length = component.length == 0 ? 1 : component.length;
+                final float length = component.length == 0 ? 1 : component.length;
                 final float ratio = componentTimer / length;
 
                 rx += component.rotChange[0] * ratio + component.rotOffset[0];
@@ -497,7 +497,7 @@ public class Animators
                 float t1 = time1, t2 = time2;
                 if (animation.loops)
                 {
-                    int l = animChannel.length();
+                    float l = animChannel.length();
                     if (l > 1)
                     {
                         t1 = time1 % l;
@@ -526,7 +526,7 @@ public class Animators
 
                 float componentTimer = time - component.startKey;
                 if (componentTimer > component.length) componentTimer = component.length;
-                final int length = component.length == 0 ? 1 : component.length;
+                final float length = component.length == 0 ? 1 : component.length;
                 final float ratio = componentTimer / length;
 
                 px += component.posChange[0] * ratio + component.posOffset[0];
@@ -543,7 +543,7 @@ public class Animators
                 float t1 = time1, t2 = time2;
                 if (animation.loops)
                 {
-                    int l = animChannel.length();
+                    float l = animChannel.length();
                     if (l > 1)
                     {
                         t1 = time1 % l;
@@ -572,7 +572,7 @@ public class Animators
 
                 float componentTimer = time - component.startKey;
                 if (componentTimer > component.length) componentTimer = component.length;
-                final int length = component.length == 0 ? 1 : component.length;
+                final float length = component.length == 0 ? 1 : component.length;
                 final float ratio = componentTimer / length;
 
                 sx *= component.scaleChange[0] * ratio + component.scaleOffset[0];
@@ -589,7 +589,7 @@ public class Animators
                 float t1 = time1, t2 = time2;
                 if (animation.loops)
                 {
-                    int l = animChannel.length();
+                    float l = animChannel.length();
                     if (l > 1)
                     {
                         t1 = time1 % l;
@@ -617,7 +617,7 @@ public class Animators
 
                 float componentTimer = time - component.startKey;
                 if (componentTimer > component.length) componentTimer = component.length;
-                final int length = component.length == 0 ? 1 : component.length;
+                final float length = component.length == 0 ? 1 : component.length;
                 final float ratio = componentTimer / length;
 
                 alpha_scale *= component.opacityOffset + ratio * component.opacityChange;
@@ -632,7 +632,7 @@ public class Animators
                 float t1 = time1, t2 = time2;
                 if (animation.loops)
                 {
-                    int l = animChannel.length();
+                    float l = animChannel.length();
                     if (l > 1)
                     {
                         t1 = time1 % l;
@@ -661,7 +661,7 @@ public class Animators
 
                 float componentTimer = time - component.startKey;
                 if (componentTimer > component.length) componentTimer = component.length;
-                final int length = component.length == 0 ? 1 : component.length;
+                final float length = component.length == 0 ? 1 : component.length;
                 final float ratio = componentTimer / length;
 
                 red_scale *= component.colChange[0] * ratio + component.colOffset[0];
@@ -682,7 +682,7 @@ public class Animators
                 float t1 = time1, t2 = time2;
                 if (animation.loops)
                 {
-                    int l = animChannel.length();
+                    float l = animChannel.length();
                     if (l > 1)
                     {
                         t1 = time1 % l;
@@ -726,7 +726,7 @@ public class Animators
         }
 
         @Override
-        public int getLength()
+        public float getLength()
         {
             return this.length;
         }
